Organize types into cluster/plan/storage type files
